+++ src/lerobot/robots/so_follower/__init__.py
@@ -19,7 +19,7 @@ from .config_so_follower import (
     SO101FollowerConfig,
     SOFollowerConfig,
     SOFollowerRobotConfig,
-    SO101FollowerDragontactilConfig,
+    SO101FollowerDragontactileConfig,
 )
 from .so_follower import SO100Follower, SO101Follower, SOFollower
