+++ src/lerobot/robots/so_follower/config_so_follower.py
@@ -54,9 +54,9 @@ class SOFollowerRobotConfig(RobotConfig, SOFollowerConfig):
 class SO101FollowerDragontactileConfig(SOFollowerRobotConfig):
     pass
 
-@RobotConfig.register_subclass("so101_follower_dragontactil")
+@RobotConfig.register_subclass("so101_follower_dragontactile")
 @dataclass
-class SO101FollowerDragontactilConfig(RobotConfig, SOFollowerConfig):
+class SO101FollowerDragontactileConfig(RobotConfig, SOFollowerConfig):
     pass
 
 SO100FollowerConfig = SOFollowerRobotConfig
